requirement.txt
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-dysmsapi==2.1.2
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
cryptography==43.0.3
distro==1.9.0
Django==4.2.3
django-cors-headers==4.6.0
django-filter==24.3
django-redis==5.4.0
django-simpleui==2024.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
jmespath==0.10.0
openai==1.52.2
pillow==10.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMySQL==1.1.1
PyYAML==6.0.2
redis==5.2.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
